% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{fn_subtable}
\alias{fn_subtable}
\title{Create a subtable for use within fn_table}
\usage{
fn_subtable(
  data,
  main,
  sub,
  stat_option = "both",
  custom_stat_cat = NULL,
  custom_stat_cont = NULL,
  digits = 0,
  digits_2 = 0
)
}
\arguments{
\item{data}{A data frame or survey design object}

\item{main}{The main variable to summarize}

\item{sub}{The variable to stratify by}

\item{stat_option}{String specifying the statistics to display}

\item{custom_stat_cat}{Custom statistic format for categorical variables}

\item{custom_stat_cont}{Custom statistic format for continuous variables}

\item{digits}{Number of digits for primary statistics}

\item{digits_2}{Number of digits for secondary statistics}
}
\value{
A gtsummary table object
}
\description{
Internal function that creates a single subtable for a given variable.
This function is not meant to be called directly by users.
}
\keyword{internal}
