% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveySuite-package.R
\docType{package}
\name{SurveySuite-package}
\alias{SurveySuite}
\alias{SurveySuite-package}
\title{SurveySuite: Make Crosstabs with Multiple X and Y Variables}
\description{
Based on some code I spent a lot of time digging up, I wanted an easy way to make a crosstab of a survey that could compare multiple demographic variables with multiple survey questions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hunterjohnson11/SurveySuite}
  \item Report bugs at \url{https://github.com/hunterjohnson11/SurveySuite/issues}
}

}
\keyword{internal}
